intermediate preservation

+++ app/src/main/java/com/example/ridelinkdashboard/domain/entiy/DashBoard.kt
@@ -45,24 +45,6 @@ import com.google.gson.annotations.SerializedName
 //        )
 //}
 //
-//{
-//        "acc_x": -0.6999806770744976,
-//        "acc_y": 0.07623741401525172,
-//        "acc_z": -0.4283081149112328,
-//        "adp_timestamp": 1620811738311,
-//        "speed": 0,
-//        "rpm": 3241,
-//        "gyro_leaning_angle": 130,
-//        "internal_battery_voltage": 6.0670046646674685,
-//        "power_battery_voltage": 4.836665837172286,
-//        "throttle_position": 33,
-//        "engine_load": 78,
-//        "temp_coolant_engine": 77,
-//        "temp_engine_oil": 3,
-//        "pressure_barometric": 96,
-//        "vin": "R1D3L1NKS1MUL4T0R"
-//}
-
 data class DashBoard(
         @SerializedName("rpm")
         val rpm: Int?,
@@ -77,7 +59,7 @@ data class DashBoard(
         @SerializedName("temp_coolant_engine")
         val temp_coolant_engine: Double?,
         @SerializedName("temp_engine_oil")
-        val temp_engine_oil: Int?,
+        val temp_engine_oil: Double?,
         @SerializedName("gyro_leaning_angle")
         val gyro_leaning_angle: Int?,
         @SerializedName("pressure_barometric")
